requirements.txt
polygon-api-client>=1.13.3
alpaca-py>=0.42.0
alpha_vantage
ibapi==9.81.1.post1
yfinance>=0.2.61
matplotlib>=3.3.3
quandl
numpy>=1.20.0
pandas>=2.2.0
polars>=1.32.3
pandas_market_calendars>=5.1.0
pandas-ta-classic>=0.3.14b0
plotly>=5.18.0
sqlalchemy
bcrypt
pytest
yappi>=1.6.0
scipy>=1.14.0
quantstats-lumi>=1.1.3,<1.2.0
python-dotenv  # Secret Storage
ccxt>=4.4.80
termcolor>=2.0
jsonpickle
apscheduler>=3.10.4
appdirs
pyarrow>=15.0.0
tqdm
lumiwealth-tradier>=0.1.18
pytz
psycopg2-binary
exchange_calendars>=4.6.0
duckdb
tabulate
databento>=0.42.0
holidays
psutil
openai
google-adk>=1.19.0
google-genai>=1.68.0
anyio>=4.10.0
mcp>=1.26.0
schwab-py>=1.5.0
Flask>=2.3
free-proxy
requests-oauthlib
boto3>=1.40.64
